pop3service/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Script==2.0.6
psycopg2-binary==2.9.11
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
gunicorn==21.2.0

# 개발 도구
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.0.0